server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate
Flask-Bcrypt
psycopg2-binary
SQLAlchemy
python-dotenv
marshmallow
pandas
openpyxl
redis
gunicorn
